function that sorts an array of integers in ascending order using the LSD Radix sort algorithm

+++ 0x1B-radix_sort/0-radix_sort.c
@@ -8,7 +8,7 @@
 int getMax(int *array, size_t size)
 {
 int mx = array[0];
-size_t i;
+uint i;
 for (i = 1; i < size; i++)
 if (array[i] > mx)
 mx = array[i];
